requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.0.5
django-admin-interface==0.16.3
django-background-tasks==1.2.5
django-colorfield==0.4.1
django-compat==1.0.15
djangorestframework==3.8.2
djangorestframework-api-key==2.0.0
djongo==1.3.4
dnspython==2.1.0
drfapikey==0.0.3
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
Jinja2==3.0.0
jwcrypto==0.8
MarkupSafe==2.0.0
numpy==1.20.3
oauth2client==4.1.3
packaging==20.9
Pillow==8.2.0
protobuf==3.17.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
Pyrebase4==4.4.3
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
sqlparse==0.2.4
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.4
